requirements.txt
backcall==0.2.0
certifi==2021.5.30
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.21.0
parso==0.7.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.1.3
prompt-toolkit==3.0.17
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pywin32==228
pyzmq==20.0.0
scikit-learn==0.24.2
scipy==1.7.0
setuptools==52.0.0.post20210125
six==1.16.0
sklearn==0.0
spyder-kernels==2.0.3
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
wheel==0.36.2
wincertstore==0.2
